+++ lib/CabalToDhall.hs
@@ -15,7 +15,7 @@ module CabalToDhall
   , getDefault
   ) where
 
-import Data.Foldable ( foldMap )
+import Data.Foldable ( foldMap, toList )
 import Data.Functor.Contravariant ( (>$<), Contravariant( contramap ) )
 import Data.Semigroup ( Semigroup, (<>) )
 import Data.Void ( absurd )
@@ -24,6 +24,7 @@ import Numeric.Natural ( Natural )
 import qualified Data.ByteString as ByteString
 import qualified Data.Sequence as Seq
 import qualified Data.Text as StrictText
+import qualified Data.Text.Encoding as StrictText
 import qualified Dhall
 import qualified Dhall.Core
 import qualified Dhall.Core as Expr ( Expr(..), Var(..), Chunks(..), makeBinding )
@@ -34,9 +35,9 @@ import qualified Distribution.Compiler as Cabal
 import qualified Distribution.License as Cabal
 import qualified Distribution.ModuleName as Cabal
 import qualified Distribution.PackageDescription.Parsec as Cabal
+import qualified Distribution.Pretty as Cabal
 import qualified Distribution.SPDX as SPDX
 import qualified Distribution.System as Cabal
-import qualified Distribution.Text as Cabal
 import qualified Distribution.Types.Benchmark as Cabal
 import qualified Distribution.Types.BenchmarkInterface as Cabal
 import qualified Distribution.Types.BuildInfo as Cabal
@@ -54,6 +55,8 @@ import qualified Distribution.Types.GenericPackageDescription as Cabal
 import qualified Distribution.Types.IncludeRenaming as Cabal
 import qualified Distribution.Types.LegacyExeDependency as Cabal
 import qualified Distribution.Types.Library as Cabal
+import qualified Distribution.Types.LibraryName as Cabal
+import qualified Distribution.Types.LibraryVisibility as Cabal
 import qualified Distribution.Types.Mixin as Cabal
 import qualified Distribution.Types.ModuleReexport as Cabal
 import qualified Distribution.Types.ModuleRenaming as Cabal
@@ -62,6 +65,8 @@ import qualified Distribution.Types.PackageId as Cabal
 import qualified Distribution.Types.PackageName as Cabal
 import qualified Distribution.Types.PkgconfigDependency as Cabal
 import qualified Distribution.Types.PkgconfigName as Cabal
+import qualified Distribution.Types.PkgconfigVersion as Cabal
+import qualified Distribution.Types.PkgconfigVersionRange as Cabal
 import qualified Distribution.Types.SetupBuildInfo as Cabal
 import qualified Distribution.Types.SourceRepo as Cabal
 import qualified Distribution.Types.TestSuite as Cabal
@@ -112,7 +117,8 @@ cabalToDhall dhallLocation genericPackageDescription
 data KnownDefault
   = CompilerOptions
   | BuildInfo
-  | Library
+  | MainLibrary
+  | NamedLibrary
   | Executable
   | Benchmark
   | TestSuite
@@ -126,6 +132,7 @@ data PreludeReference
   | PreludeConstructorsLicense
   | PreludeConstructorsRepoKind
   | PreludeConstructorsScope
+  | PreludeConstructorsLibraryVisibility
   | PreludeV
 
 
@@ -141,6 +148,8 @@ resolvePreludeVar = \case
     Expr.Var "types" `Expr.Field` "RepoKind"
   PreludeConstructorsScope ->
     Expr.Var "types" `Expr.Field` "Scope"
+  PreludeConstructorsLibraryVisibility ->
+    Expr.Var "types" `Expr.Field` "LibraryVisibility"
 
 
 type Default s a
@@ -178,8 +187,10 @@ getDefault typesLoc resolve typ = withTypesImport expr
           Expr.RecordLit ( compilerOptionsDefault resolve )
         BuildInfo ->
           Expr.RecordLit ( buildInfoDefault resolve )
-        Library ->
-          factorBuildInfo ( libraryDefault resolve )
+        MainLibrary ->
+          factorBuildInfo ( libraryDefault False resolve )
+        NamedLibrary ->
+          factorBuildInfo ( libraryDefault True resolve )
         Executable ->
           factorBuildInfo ( executableDefault resolve )
         Benchmark ->
@@ -225,17 +236,8 @@ generaliseDeclared =
 compilerOptionsDefault :: Default s a
 compilerOptionsDefault _resolve =
   ( Map.fromList
-    [ emptyListDefault "Eta" Expr.Text
-    , emptyListDefault "GHC" Expr.Text
+    [ emptyListDefault "GHC" Expr.Text
     , emptyListDefault "GHCJS" Expr.Text
-    , emptyListDefault "HBC" Expr.Text
-    , emptyListDefault "Helium" Expr.Text
-    , emptyListDefault "Hugs" Expr.Text
-    , emptyListDefault "JHC" Expr.Text
-    , emptyListDefault "LHC" Expr.Text
-    , emptyListDefault "NHC" Expr.Text
-    , emptyListDefault "UHC" Expr.Text
-    , emptyListDefault "YHC" Expr.Text
     ]
   )
 
@@ -284,11 +286,26 @@ buildInfoDefault resolve = fields
       , emptyListDefault "virtual-modules" Expr.Text
       , emptyListDefault "extra-lib-flavours" Expr.Text
       , emptyListDefault "extra-bundled-libs" Expr.Text
+      , emptyListDefault "autogen-includes" Expr.Text
+      , emptyListDefault "extra-dyn-lib-flavours" Expr.Text
       ]
 
 
-libraryDefault :: Default s a
-libraryDefault resolve = buildInfoDefault resolve <> specificFields
+libraryVisibility :: Dhall.InputType Cabal.LibraryVisibility
+libraryVisibility =
+  Dhall.InputType
+    { Dhall.embed = \case
+        Cabal.LibraryVisibilityPublic ->
+            Expr.Var "types" `Expr.Field` "LibraryVisibility" `Expr.Field` "public"
+        Cabal.LibraryVisibilityPrivate ->
+            Expr.Var "types" `Expr.Field` "LibraryVisibility" `Expr.Field` "private"
+    , Dhall.declared =
+        Expr.Var "types" `Expr.Field` "LibraryVisibility"
+    }
+
+
+libraryDefault :: Bool -> Default s a
+libraryDefault named resolve = buildInfoDefault resolve <> specificFields
   where
     specificFields = Map.fromList
       [ emptyListDefault "exposed-modules" Expr.Text
@@ -296,7 +313,13 @@ libraryDefault resolve = buildInfoDefault resolve <> specificFields
       , emptyListDefault "reexported-modules"
           ( generaliseDeclared moduleReexport )
       , emptyListDefault "signatures" Expr.Text
+      , ( "visibility"
+        , ( resolve PreludeConstructorsLibraryVisibility `Expr.Field` visibility )
+        )
       ]
+    visibility = if named
+      then "private"
+      else "public"
 
 
 benchmarkDefault :: Default s a
@@ -359,7 +382,7 @@ packageDefault resolve = fields
           ( Expr.Pi
               "config"
               ( Expr.Var "types" `Expr.Field` "Config" )
-              ( generaliseDeclared library )
+              ( generaliseDeclared ( library False ) )
           )
       , ( "license"
         , resolve PreludeConstructorsLicense `Expr.Field` "AllRightsReserved"
@@ -369,7 +392,7 @@ packageDefault resolve = fields
       , textFieldDefault "package-url" ""
       , emptyListDefault "source-repos" ( generaliseDeclared sourceRepo )
       , textFieldDefault "stability" ""
-      , emptyListDefault "sub-libraries" ( named "library" library )
+      , emptyListDefault "sub-libraries" ( named "library" ( library True ) )
       , textFieldDefault "synopsis" ""
       , emptyListDefault "test-suites" ( named "test-suite" testSuite )
       , emptyListDefault "tested-with"
@@ -512,8 +535,8 @@ genericPackageDescriptionToDhall =
     ( mconcat
         [ Cabal.packageDescription >$< packageDescriptionToRecord
         , recordField "flags" ( Cabal.genPackageFlags >$< ( listOf flag ) )
-        , recordField "library" ( Cabal.condLibrary >$< maybeToDhall ( condTree library ) )
-        , recordField "sub-libraries" ( Cabal.condSubLibraries >$< named "library" ( condTree library ) )
+        , recordField "library" ( Cabal.condLibrary >$< maybeToDhall ( condTree ( library False ) ) )
+        , recordField "sub-libraries" ( Cabal.condSubLibraries >$< named "library" ( condTree ( library True ) ) )
         , recordField "foreign-libraries" ( Cabal.condForeignLibs >$< named "foreign-lib" ( condTree foreignLibrary ) )
         , recordField "executables" ( Cabal.condExecutables >$< named "executable" ( condTree executable ) )
         , recordField "test-suites" ( Cabal.condTestSuites >$< named "test-suite" ( condTree testSuite ) )
@@ -586,7 +609,7 @@ versionToDhall =
         Expr.App ( Expr.Var "prelude" `Expr.Field` "v" )
           . Dhall.embed stringToDhall
           . show
-          . Cabal.disp
+          . Cabal.pretty
     , Dhall.declared =
         Expr.Var "types" `Expr.Field` "Version"
     }
@@ -985,14 +1008,34 @@ setupBuildInfo =
     }
 
 
+libraryName :: Dhall.InputType Cabal.LibraryName
+libraryName =
+  Dhall.InputType
+    { Dhall.embed = \case
+        Cabal.LMainLibName ->
+          Expr.Var "types" `Expr.Field` "LibraryName" `Expr.Field` "main-library"
+        Cabal.LSubLibName c ->
+          Expr.App
+            ( Expr.Var "types" `Expr.Field` "LibraryName" `Expr.Field` "sub-library" )
+            ( Dhall.embed unqualComponentName c )
+    , Dhall.declared =
+        Expr.Var "types" `Expr.Field` "LibraryName"
+    }
+
+
 dependency :: Dhall.InputType Cabal.Dependency
 dependency =
-  runRecordInputType
+  ( runRecordInputType
     ( mconcat
-        [ recordField "package" ( contramap ( \( Cabal.Dependency p _ ) -> p ) packageNameToDhall )
-        , recordField "bounds" ( contramap ( \( Cabal.Dependency _ a ) -> a ) versionRange )
+        [ recordField "package" ( contramap ( \( Cabal.Dependency p _ _ ) -> p ) packageNameToDhall )
+        , recordField "bounds" ( contramap ( \( Cabal.Dependency _ a _ ) -> a ) versionRange )
+        , recordField "library-names" ( contramap ( \( Cabal.Dependency _ _ ls ) -> toList ls ) ( listOf libraryName ) )
         ]
     )
+  )
+    { Dhall.declared =
+        Expr.Var "types" `Expr.Field` "Dependency"
+    }
 
 
 flag :: Dhall.InputType Cabal.Flag
@@ -1012,9 +1055,11 @@ flagName =
   contramap Cabal.unFlagName stringToDhall
 
 
-library :: Dhall.InputType Cabal.Library
-library =
-  ( runRecordInputTypeWithDefault Library libraryDefault
+library :: Bool -> Dhall.InputType Cabal.Library
+library named =
+  ( runRecordInputTypeWithDefault
+      ( if named then NamedLibrary else MainLibrary )
+      ( libraryDefault named )
       ( mconcat
           [ contramap Cabal.libBuildInfo buildInfoRecord
           , recordField
@@ -1026,6 +1071,9 @@ library =
           , recordField
               "signatures"
               ( contramap Cabal.signatures ( listOf moduleName ) )
+          , recordField
+              "visibility"
+              ( contramap Cabal.libVisibility libraryVisibility )
           ]
       )
   )
@@ -1089,7 +1137,7 @@ condTree t =
 
 moduleName :: Dhall.InputType Cabal.ModuleName
 moduleName =
-  contramap ( show . Cabal.disp ) stringToDhall
+  contramap ( show . Cabal.pretty ) stringToDhall
 
 
 condBranchCondition :: Dhall.InputType (Cabal.Condition Cabal.ConfVar)
@@ -1279,6 +1327,8 @@ buildInfoRecord =
     , recordField "virtual-modules" ( Cabal.virtualModules >$< listOf moduleName )
     , recordField "extra-lib-flavours" ( Cabal.extraLibFlavours >$< listOf stringToDhall )
     , recordField "extra-bundled-libs" ( Cabal.extraBundledLibs >$< listOf stringToDhall )
+    , recordField "autogen-includes" ( Cabal.autogenIncludes >$< listOf stringToDhall )
+    , recordField "extra-dyn-lib-flavours" ( Cabal.extraDynLibFlavours >$< listOf stringToDhall )
     ]
 
 
@@ -1322,7 +1372,7 @@ exeDependency =
 
 unqualComponentName :: Dhall.InputType Cabal.UnqualComponentName
 unqualComponentName =
-  show . Cabal.disp >$< stringToDhall
+  show . Cabal.pretty >$< stringToDhall
 
 
 pkgconfigDependency :: Dhall.InputType Cabal.PkgconfigDependency
@@ -1330,14 +1380,66 @@ pkgconfigDependency =
   runRecordInputType
     ( mconcat
         [ recordField "name" ( ( \( Cabal.PkgconfigDependency a _version ) -> a ) >$< pkgconfigName )
-        , recordField "version" ( ( \( Cabal.PkgconfigDependency _name a ) -> a ) >$< versionRange )
+        , recordField "version" ( ( \( Cabal.PkgconfigDependency _name a ) -> a ) >$< pkgconfigVersionRange )
         ]
     )
 
 
 pkgconfigName :: Dhall.InputType Cabal.PkgconfigName
 pkgconfigName =
-  show . Cabal.disp >$< stringToDhall
+  show . Cabal.pretty >$< stringToDhall
+
+
+-- PkgconfigVersion is restricted to ASCII-only characters.
+pkgconfigVersion :: Dhall.InputType Cabal.PkgconfigVersion
+pkgconfigVersion = (\ ( Cabal.PkgconfigVersion a ) -> StrictText.decodeLatin1 a ) >$< Dhall.inject
+
+pkgconfigVersionRange :: Dhall.InputType Cabal.PkgconfigVersionRange
+pkgconfigVersionRange =  Dhall.InputType
+    { Dhall.embed =
+          let
+            go = \case
+              Cabal.PcAnyVersion ->
+                Expr.Var "prelude" `Expr.Field` "pkg-config" `Expr.Field` "anyVersion"
+              Cabal.PcThisVersion v ->
+                Expr.App
+                  ( Expr.Var "prelude" `Expr.Field` "pkg-config" `Expr.Field` "thisVersion" )
+                  ( Dhall.embed pkgconfigVersion v )
+              Cabal.PcLaterVersion v ->
+                Expr.App
+                  ( Expr.Var "prelude" `Expr.Field` "pkg-config" `Expr.Field` "laterVersion" )
+                  ( Dhall.embed pkgconfigVersion v )
+              Cabal.PcEarlierVersion v ->
+                Expr.App
+                  ( Expr.Var "prelude" `Expr.Field` "pkg-config" `Expr.Field` "earlierVersion" )
+                  ( Dhall.embed pkgconfigVersion v )
+              Cabal.PcOrLaterVersion v ->
+                Expr.App
+                  ( Expr.Var "prelude" `Expr.Field` "pkg-config" `Expr.Field` "orLaterVersion" )
+                  ( Dhall.embed pkgconfigVersion v )
+              Cabal.PcOrEarlierVersion v ->
+                Expr.App
+                  ( Expr.Var "prelude" `Expr.Field` "pkg-config" `Expr.Field` "orEarlierVersion" )
+                  ( Dhall.embed pkgconfigVersion v )
+              Cabal.PcUnionVersionRanges a b ->
+                Expr.App
+                  ( Expr.App
+                    ( Expr.Var "prelude" `Expr.Field` "pkg-config" `Expr.Field` "unionVersionRanges" )
+                    ( go a )
+                  )
+                  ( go b )
+              Cabal.PcIntersectVersionRanges a b ->
+                Expr.App
+                  ( Expr.App
+                    ( Expr.Var "prelude" `Expr.Field` "pkg-config" `Expr.Field` "intersectVersionRanges" )
+                    ( go a )
+                  )
+                  ( go b )
+          in
+          go
+    , Dhall.declared =
+        Expr.Var "types" `Expr.Field` "PkgconfigVersionRange"
+    }
 
 
 language :: Dhall.InputType Cabal.Language
@@ -1389,24 +1491,16 @@ extension =
       ( Expr.BoolLit trueFalse )
 
 
-compilerOptions :: Dhall.InputType [ ( Cabal.CompilerFlavor, [ String ] ) ]
+compilerOptions :: Dhall.InputType ( Cabal.PerCompilerFlavor [ String ] )
 compilerOptions =
-  Dhall.InputType
-    { Dhall.embed = \xs ->
-        withDefault CompilerOptions compilerOptionsDefault
-          ( Expr.RecordLit
-              ( Map.fromList
-                  ( map
-                      ( \( c, opts ) ->
-                          ( StrictText.pack ( show c )
-                          , Expr.ListLit ( Just ( Expr.App Expr.List Expr.Text ) ) ( dhallString <$> Seq.fromList opts )
-                          )
-                      )
-                      xs
-                  )
-              )
-          )
-    , Dhall.declared =
+  ( runRecordInputTypeWithDefault CompilerOptions compilerOptionsDefault
+      ( mconcat
+          [ recordField "GHC" ( (\ ( Cabal.PerCompilerFlavor ghc _ ) -> ghc ) >$< listOf stringToDhall )
+          , recordField "GHCJS" ( (\ ( Cabal.PerCompilerFlavor _ ghcjs ) -> ghcjs ) >$< listOf stringToDhall )
+          ]
+      )
+  )
+    { Dhall.declared =
         Expr.Var "types" `Expr.Field` "CompilerOptions"
     }