+++ overrides/ghc-paths.nix
@@ -1,0 +1,10 @@
+{ mkDerivation, base, Cabal, directory, stdenv }:
+mkDerivation {
+  pname = "ghc-paths";
+  version = "0.1.0.12";
+  sha256 = "6ecbe676d073cb07989c61ce4c5709c4e67cbefdd2d55a4095f9388b6fe2c484";
+  setupHaskellDepends = [ base Cabal directory ];
+  libraryHaskellDepends = [ base ];
+  description = "Knowledge of GHC's installation directories";
+  license = stdenv.lib.licenses.bsd3;
+}